Extract Keyboard Accessory to its own class
The code to add a : to the keyboard was duplicated in two controllers. To
make it easier to change this functionality the shared code is extracted
to a component

Test Plan:
* Verified the code compiles
* Checked the : still appears in both locations and tapping it works
