+++ BeeSwift/EditDatapointViewController.swift
@@ -72,24 +72,9 @@ class EditDatapointViewController: UIViewController, UITextFieldDelegate {
         self.valueField.text = "\(self.datapointJSON!["value"].number!)"
         self.valueField.addTarget(self, action: #selector(self.textFieldEditingDidBegin), for: .editingDidBegin)
         
-        let accessory = UIView(frame: CGRect(x: 0, y: 0, width: 0, height: 44))
-        accessory.backgroundColor = UIColor.white
+        let accessory = DatapointValueAccessory()
+        accessory.valueField = self.valueField
         self.valueField.inputAccessoryView = accessory
-        let colonButton = UIButton()
-        accessory.addSubview(colonButton)
-        accessory.clipsToBounds = true
-        colonButton.snp.makeConstraints { (make) -> Void in
-            make.width.equalTo(accessory).multipliedBy(1.0/3.0).offset(-1)
-            make.height.equalTo(accessory)
-            make.left.equalTo(-1)
-            make.top.equalTo(0)
-        }
-        colonButton.setTitle(":", for: UIControl.State())
-        colonButton.layer.borderWidth = 1
-        colonButton.layer.borderColor = UIColor.beeminder.gray.cgColor
-        colonButton.titleLabel?.font = UIFont.boldSystemFont(ofSize: 26)
-        colonButton.setTitleColor(UIColor.black, for: UIControl.State())
-        colonButton.addTarget(self, action: #selector(self.colonButtonPressed), for: .touchUpInside)
         
         self.scrollView.addSubview(self.commentField)
         self.commentField.snp.makeConstraints { (make) in
@@ -147,10 +132,6 @@ class EditDatapointViewController: UIViewController, UITextFieldDelegate {
             make.height.equalTo(0)
         }
         self.datePicker.isHidden = true
-    }
-    
-    @objc func colonButtonPressed() {
-        self.valueField.text = "\(self.valueField.text!):"
     }
     
     func textField(_ textField: UITextField, shouldChangeCharactersIn range: NSRange, replacementString string: String) -> Bool {
